+++ lib/shadcn/components/menu/menu.dart
@@ -481,7 +481,7 @@ class _MenuButtonState extends State<MenuButton> {
               }),
         ),
       ),
-    );
+    ).blurIn;
   }
 }